+++ scripts/discover_sgns.sh
@@ -10,13 +10,15 @@ min_count2=$7
 itera=$8
 t=$9
 language=${10}
-embed_pretrained=${11}
-path_pretrained=${12}
-sample_id=${13}
-sample_size=${14}
-max_usages=${15}
+incremental_learning=${11}
+w2vec_algorithm=${12}
+embed_pretrained=${13}
+path_pretrained=${14}
 
-max_samples=${16}
+sample_id=${15}
+sample_size=${16}
+max_usages=${17}
+max_samples=${18}
 
 function usage {
     echo "Given a corpus pair C1 and C2, decide for the intersection of their vocabularies which words lost or gained sense(s) between C1 and C2."
@@ -27,30 +29,33 @@ function usage {
     echo "      discover_sgns.sh <data_set_id> <window_size> <dim> <k> <s> <min_count> <itera> <t> <language> <sample_id> <sample_size> max_usages>" 
     echo "      discover_sgns.sh <data_set_id> <window_size> <dim> <k> <s> <min_count> <itera> <t> <language> <sample_id> <sample_size> <max_usages> <max_samples>" 
     echo ""
-    echo "      <data_set_id>       = data set param_id"
-    echo "      <window_size>       = the linear distance of context words to consider in each direction"
-    echo "      <dim>               = dimensionality of embeddings"
-    echo "      <k>                 = number of negative samples parameter (equivalent to shifting parameter for PPMI)"
-    echo "      <s>                 = threshold for subsampling"
-    echo "      <min_count1>        = number of occurrences for a word to be included in the vocabulary (corpus1)"
-    echo "      <min_count2>        = number of occurrences for a word to be included in the vocabulary (corpus2)"
-    echo "      <itera>             = number of iterations"
-    echo "      <t>                 = threshold = mean + t * standard deviation"
-    echo "      <language>          = en | de"
-    echo "      <embed_pretrained>  = option of pretrained embeddings (None, Glove)"
-    echo "      <path_pretrained>   = path to pretrained embeddings directory with txt files"
-    echo "      <sample_id>         = sample identifier"    
-    echo "      <sample_size>       = Number of words to be sampled from filtered words (after filter1)"
-    echo "      <max_usages>        = max. number of usages to be extracted from each corpus"
-    echo "      <max_samples>       = max. number of samples stored for annotation"
+    echo "      <data_set_id>           = data set param_id"
+    echo "      <window_size>           = the linear distance of context words to consider in each direction"
+    echo "      <dim>                   = dimensionality of embeddings"
+    echo "      <k>                     = number of negative samples parameter (equivalent to shifting parameter for PPMI)"
+    echo "      <s>                     = threshold for subsampling"
+    echo "      <min_count1>            = number of occurrences for a word to be included in the vocabulary (corpus1)"
+    echo "      <min_count2>            = number of occurrences for a word to be included in the vocabulary (corpus2)"
+    echo "      <itera>                 = number of iterations"
+    echo "      <t>                     = threshold = mean + t * standard deviation"
+    echo "      <language>              = en | de"
+    echo "      <incremental_learning>  = indicates incremental learning (incr,nonincr)"
+    echo "      <w2vec_algorithm>       = cbow | sgns"
+    echo "      <embed_pretrained>      = option of pretrained embeddings (None, Glove)"
+    echo "      <path_pretrained>       = path to pretrained embeddings directory with txt files"
+    echo "      <sample_id>             = sample identifier"
+    echo "      <sample_size>           = Number of words to be sampled from filtered words (after filter1)"
+    echo "      <max_usages>            = max. number of usages to be extracted from each corpus"
+    echo "      <max_samples>           = max. number of samples stored for annotation"
     echo ""
 }
 
-if [ $# -ne 10 ]  && [ $# -ne 12 ] && [ $# -ne 15 ] && [ $# -ne 16 ]
+if [ $# -ne 10 ]  && [ $# -ne 11 ] && [ $# -ne 12 ] && [ $# -ne 14 ]&& [ $# -ne 17 ] && [ $# -ne 18 ]
 	then 
 		usage
 		exit 1
 fi
+
 
 if [[ ( $1 == "--help") ||  $1 == "-h" ]] 
 	then 
@@ -58,33 +63,39 @@ if [[ ( $1 == "--help") ||  $1 == "-h" ]]
 		exit 0
 fi
 
-param_id=SGNS_win${window_size}_dim${dim}_k${k}_s${s}_mc${min_count1}_mc${min_count2}_i${itera}
+param_id=${w2vec_algorithm}_win${window_size}_dim${dim}_k${k}_s${s}_mc${min_count1}_mc${min_count2}_i${itera}_${incremental_learning}
 
-
-if [ $# -eq 12 ]
+if [ $# -eq 14 ]
   then
     param_id+=_${embed_pretrained}
 fi
 
-
 outdir=output/${data_set_id}/${param_id}/discovery/t${t}
 resdir=results/${data_set_id}/${param_id}/discovery/t${t}
-targetdir=data/${data_set_id}/targets
 
 mkdir -p ${outdir}
 mkdir -p ${resdir}
 
 
 # Generate static word embeddings with SGNS
-python static/sgns.py ./data/${data_set_id}/corpus1/lemma.txt.gz ${outdir}/mat1 ${window_size} ${dim} ${k} ${s} ${min_count1} ${itera} ${embed_pretrained} ${path_pretrained}
-python static/sgns.py ./data/${data_set_id}/corpus2/lemma.txt.gz ${outdir}/mat2 ${window_size} ${dim} ${k} ${s} ${min_count2} ${itera} ${embed_pretrained} ${path_pretrained}
+python static/sgns.py ./data/${data_set_id}/corpus1/lemma.txt.gz ${outdir}/mat1 ${window_size} ${dim} ${k} ${s} ${min_count1} ${itera} ${incremental_learning} ${w2vec_algorithm} --embed_pretrained ${embed_pretrained} --path_pretrained ${path_pretrained}
+python static/sgns.py ./data/${data_set_id}/corpus2/lemma.txt.gz ${outdir}/mat2 ${window_size} ${dim} ${k} ${s} ${min_count2} ${itera} ${incremental_learning} ${w2vec_algorithm} --embed_pretrained ${embed_pretrained} --path_pretrained ${path_pretrained}
 
-# Length-normalize, meanc-center and align with OP
-python modules/map_embeddings.py --normalize unit center --init_identical --orthogonal ${outdir}/mat1 ${outdir}/mat2 ${outdir}/mat1ca ${outdir}/mat2ca
+if [ "$incremental_learning" == "nonincr" ]
+  then
+    echo "Non incremental learning"
+    # Length-normalize, meanc-center and align with OP
+    python modules/map_embeddings.py --normalize unit center --init_identical --orthogonal ${outdir}/mat1 ${outdir}/mat2 ${outdir}/mat1ca ${outdir}/mat2ca
+    # Measure CD for every word in the intersection of the vocabularies
+    python measures/cd.py ${outdir}/mat1ca ${outdir}/mat2ca ${resdir}/distances_intersection.tsv
 
+elif [ "$incremental_learning" == "incr" ]
+  then
+    echo "Mapping is not necessary in cases of incremental_learning"
+    # Measure CD for every word in the intersection of the vocabularies
+    python measures/cd.py ${outdir}/mat1 ${outdir}/mat2 ${resdir}/distances_intersection.tsv
+fi
 
-# Measure CD for every word in the intersection of the vocabularies
-python measures/cd.py ${outdir}/mat1ca ${outdir}/mat2ca ${resdir}/distances_intersection.tsv
 
 
 # Create predictions
@@ -95,7 +106,7 @@ python modules/filter1.py ${resdir}/predictions.tsv ${resdir}/predictions_f1.tsv
 
 
 # Sample from <predictions_f1.tsv> 
-if [ $# -eq 13 ] || [ $# -eq 14 ]
+if [ $# -eq 17 ] || [ $# -eq 18 ]
     then
         sample=${resdir}/predictions_f1.tsv
         
@@ -130,7 +141,7 @@ fi
 
 
 # Store in DURel format
-if [ $# -eq 14 ]
+if [ $# -eq 18 ]
     then
         mkdir -p ${resdir}/${sample_id}/DURel
         number_lines=$(wc -l ${resdir}/${sample_id}/predictions_f2.tsv | awk '{print $1 }')